[Add] authentication jwt (user/mod/admin).

+++ src/controllers/userController.ts
@@ -2,6 +2,7 @@ import { Response, Request } from 'express';
 import { PrismaClient, Role } from '@prisma/client';
 import { z } from 'zod';
 import { hashPassword, verifyPassword } from '../utils/bcryptUtil';
+import validateUserIdentity from '../services/tokenJwtService/validateUserIdentity';
 
 const prisma = new PrismaClient();
 
@@ -23,7 +24,7 @@ const integerValidator = z
  * @param {Response} res - Express Response object.
  * @returns {Promise<void>}
  */
-export const getUsers = async (res: Response): Promise<void> => {
+export const getUsers = async (_: Request, res: Response): Promise<void> => {
     try {
         const users = await prisma.user.findMany({
             select: {
@@ -41,7 +42,11 @@ export const getUsers = async (res: Response): Promise<void> => {
         });
         res.status(200).json(users);
     } catch (error) {
-        res.status(500).json({ message: 'Internal server error' });
+        if (error.name === 'AdminOnly') {
+            res.status(401).json({ message: error.message });
+        } else {
+            res.status(500).json({ message: error.message || 'Internal server error' });
+        }
     }
 };
 
@@ -54,10 +59,15 @@ export const getUsers = async (res: Response): Promise<void> => {
  */
 export const createUser = async (req: Request, res: Response): Promise<void> => {
     const createUserSchema = z.object({
-        name: z.string(),
-        email: z.string().email(),
-        password: z.string(),
-        role: z.enum([Role.ADMIN, Role.USER]),
+        name: z.string().min(1).max(255),
+        email: z
+            .string()
+            .email()
+            .min(1)
+            .max(255)
+            .refine((val) => val.endsWith('@ufpr.br' || '@inf.ufpr.br')),
+        password: z.string().min(8).max(255),
+        role: z.enum([Role.ADMIN, Role.USER, Role.MOD]),
         status: z.boolean(),
         courseId: z.number(),
     });
@@ -76,7 +86,7 @@ export const createUser = async (req: Request, res: Response): Promise<void> => 
                 password,
                 role,
                 status,
-                courseId
+                courseId,
             },
         });
 
@@ -103,29 +113,39 @@ export const createUser = async (req: Request, res: Response): Promise<void> => 
  * @returns {Promise<void>}
  */
 export const updateUser = async (req: Request, res: Response): Promise<void> => {
-    const updateUserSchema = z.object({
-        newName: z.string().min(1).max(255).optional(),
-        newRole: z.enum([Role.ADMIN, Role.USER]),
-        email: z.string().email().optional(),
-        newEmail: z.string().email().optional(),
-        newCourse: z.string().optional(),
-        password: z.string().min(1).max(255).optional(),
-        newPassword: z.string().min(1).max(255).optional(),
-    }).refine((data) => 
-        (data.email && !data.newEmail) || (!data.email && data.newEmail), {
-            message: 'Email cannot be updated as some data is missing',
-            path: ['email', 'newEmail']
-        }
-    ).refine((data) =>
-        (data.password && !data.newPassword) || (!data.password && data.newPassword), {
-            message: 'Password cannot be updated as some data is missing',
-            path: ['password', 'password']
-        }
-    );
-
     try {
-        const id = await integerValidator.parseAsync(req.body.tagId);
-        const { newName, email, newEmail, newCourse, password, newPassword, newRole } = await updateUserSchema.parseAsync(req.body);
+        const updateUserSchema = z
+            .object({
+                newName: z.string().min(1).max(255).optional(),
+                email: z
+                    .string()
+                    .email()
+                    .refine((val) => val.endsWith('@ufpr.br'))
+                    .optional(),
+                newEmail: z
+                    .string()
+                    .email()
+                    .refine((val) => val.endsWith('@ufpr.br'))
+                    .optional(),
+                newCourseId: z.number().int().optional(),
+                password: z.string().min(8).max(255).optional(),
+                newPassword: z.string().min(8).max(255).optional(),
+            })
+            .refine(
+                (data) => {
+                    if ((!data.email && data.newEmail) || (data.email && !data.newEmail)) return false;
+                    if ((data.password && !data.newPassword) || (!data.password && data.newPassword)) return false;
+
+                    return true;
+                },
+                {
+                    message: 'Email or password cannot be updated as some data is missing',
+                }
+            );
+
+        const id = await integerValidator.parseAsync(req.params.userId);
+
+        const { newName, email, newEmail, newCourseId, password, newPassword } = await updateUserSchema.parseAsync(req.body);
 
         const user = await prisma.user.findUniqueOrThrow({
             where: {
@@ -133,25 +153,29 @@ export const updateUser = async (req: Request, res: Response): Promise<void> => 
             },
         });
 
-        if(email && email !== user.email)
-            throw new Error('Current email is wrong.');
-            //res.status(400).json({ message: 'Current email is wrong!'})
-        
-        const resultComparison = await verifyPassword(password, user.password);
-        if(password && !resultComparison)
-            throw new Error('Current password is wrong.');
-            //res.status(400).json({ message: 'Current password is wrong!'})
+        // validating if the target user of the update is the same as the token
+        if (!validateUserIdentity(user.email, req.headers.authorization)) throw new Error('Unauthorized user');
+
+        if (email && email !== user.email) throw new Error('Current email or password is wrong.');
+
+        if (password) {
+            const resultComparison = await verifyPassword(password, user.password);
+            if (!resultComparison) throw new Error('Current email or password is wrong.');
+        }
+
+        const encriptedNewPassword = await hashPassword(newPassword, 11);
+        if (!encriptedNewPassword) throw new Error('Unable to update password!');
 
         await prisma.user.update({
             where: {
-                id
+                id,
             },
-            data:{
+            data: {
                 name: newName,
                 email: newEmail,
-                password: newPassword,
-                role: newRole
-            }
+                password: encriptedNewPassword,
+                courseId: newCourseId,
+            },
         });
 
         res.status(200).json({ message: 'User updated!' });
@@ -159,12 +183,12 @@ export const updateUser = async (req: Request, res: Response): Promise<void> => 
         if (error.name === 'ZodError') {
             res.status(400).json({ error: error });
         } else if (error.code === 'P2025') {
-            res.status(404).json({ message: 'User not found!' });
+            res.status(400).json({ message: 'Unable to update user!' });
         } else {
-            res.status(500).json({ message: 'Internal server error' });
+            res.status(500).json({ message: error.message || 'Unable to update user!' });
         }
     }
-}
+};
 
 /**
  * Delete a user.
@@ -175,11 +199,20 @@ export const updateUser = async (req: Request, res: Response): Promise<void> => 
  */
 export const deleteUser = async (req: Request, res: Response): Promise<void> => {
     try {
-        const id = await integerValidator.parseAsync(req.body.tagId);
+        const userId = await integerValidator.parseAsync(req.params.userId);
+
+        const user = await prisma.user.findFirstOrThrow({
+            where: {
+                id: userId,
+            },
+        });
+
+        // validating if the target user of the delete is the same as the token
+        if (!validateUserIdentity(user.email, req.headers.authorization)) throw new Error();
 
         await prisma.user.delete({
             where: {
-                id: id,
+                id: userId,
             },
         });
 
@@ -187,10 +220,8 @@ export const deleteUser = async (req: Request, res: Response): Promise<void> => 
     } catch (error) {
         if (error.name === 'ZodError') {
             res.status(400).json({ error: error });
-        } else if (error.code === 'P2025') {
-            res.status(404).json({ message: 'User not found!' });
         } else {
-            res.status(500).json({ message: 'Internal server error' });
+            res.status(500).json({ message: 'Unable to delete user!' });
         }
     }
 };
